addressing butter-senpai's comments
