+++ src/services/compiler/utilites/utils.go
@@ -1,0 +1,14 @@
+package utilites
+
+import (
+	"os"
+)
+
+// Gets the value associated with the environment variable key
+// If it doesn't exist, then it just returns the fallback
+func GetEnv(key string, fallback string) string {
+	if value, ok := os.LookupEnv(key); ok {
+		return value
+	}
+	return fallback
+}
